docs: comments for DefaultActionFilter and DefaultTypeFilter
