+++ src/JsonApiDotNetCore/Middleware/DefaultExceptionFilter.cs
@@ -1,21 +1,13 @@
-using System;
-using JsonApiDotNetCore.Internal;
+﻿using JsonApiDotNetCore.Internal;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Logging;
 
 namespace JsonApiDotNetCore.Middleware
 {
-    public interface IJsonApiExceptionFilterProvider
-    {
-        Type Get();
-    }
-
-    public class JsonApiExceptionFilterProvider : IJsonApiExceptionFilterProvider
-    {
-        public Type Get() => typeof(DefaultExceptionFilter);
-    }
-
+    /// <summary>
+    /// Global exception filter that wraps any thrown error with a JsonApiException.
+    /// </summary>
     public class DefaultExceptionFilter : ActionFilterAttribute, IExceptionFilter
     {
         private readonly ILogger _logger;
